requirements.txt
# ---- Core scientific stack ----
numpy>=1.24.0
pandas>=2.0.0
polars==1.21.0
pyarrow==19.0.1

# ---- Machine learning ----
lightgbm>=4.3.0
scikit-learn>=1.4.0

# ---- Plotting & debugging ----
matplotlib>=3.8.0
seaborn>=0.13.0

# ---- Performance / convenience ----
tqdm>=4.66.0
joblib>=1.4.0

# ---- Config / parsing ----
pyyaml
python-dotenv

# ---- Cloud & filesystem ----
fsspec>=2024.2.0
adlfs>=2024.4.0
